+++ test/td_dd_web/controllers/user_search_filter_controller_test.exs
@@ -1,8 +1,6 @@
 defmodule TdDdWeb.UserSearchFilterControllerTest do
   use TdDdWeb.ConnCase
 
-  import TdDd.TestOperators
-
   @create_attrs %{
     filters: %{country: ["Spa"]},
     name: "some name",
@@ -22,8 +20,8 @@ defmodule TdDdWeb.UserSearchFilterControllerTest do
 
     @tag authentication: [role: "admin"]
     test "lists all user_search_filters filtered by scope", %{conn: conn} do
-      %{id: id} = insert(:user_search_filter, scope: :rule)
-      insert(:user_search_filter, scope: :rule_implementation)
+      %{id: id} = insert(:user_search_filter, scope: "rule")
+      insert(:user_search_filter, scope: "rule_implementation")
 
       assert %{"data" => [%{"id" => ^id}]} =
                conn
@@ -53,8 +51,8 @@ defmodule TdDdWeb.UserSearchFilterControllerTest do
       conn: conn,
       claims: %{user_id: user_id}
     } do
-      %{id: id} = insert(:user_search_filter, name: "a", user_id: user_id, scope: :rule)
-      insert(:user_search_filter, name: "b", user_id: user_id, scope: :rule_implementation)
+      %{id: id} = insert(:user_search_filter, name: "a", user_id: user_id, scope: "rule")
+      insert(:user_search_filter, name: "b", user_id: user_id, scope: "rule_implementation")
 
       assert %{"data" => data} =
                conn
@@ -64,21 +62,21 @@ defmodule TdDdWeb.UserSearchFilterControllerTest do
       assert [%{"id" => ^id}] = data
     end
 
-    @tag authentication: [user_name: "non_admin"]
+    @tag authentication: [user_name: "non_admin", permissions: ["view_quality_rule"]]
     test "lists current user user_search_filters with global filters", %{
       conn: conn,
       claims: %{user_id: user_id}
     } do
-      %{id: id1} = insert(:user_search_filter, name: "a", user_id: user_id)
-      %{id: id2} = insert(:user_search_filter, name: "b", is_global: true)
-      insert(:user_search_filter, name: "c", is_global: false)
+      %{id: id1} = insert(:user_search_filter, scope: "rule", name: "a", user_id: user_id)
+      %{id: id2} = insert(:user_search_filter, scope: "rule", name: "b", is_global: true)
+      insert(:user_search_filter, scope: "rule", name: "c", is_global: false)
 
       assert %{"data" => data} =
                conn
-               |> get(Routes.user_search_filter_path(conn, :index_by_user))
+               |> get(Routes.user_search_filter_path(conn, :index_by_user), %{"scope" => "rule"})
                |> json_response(:ok)
 
-      assert [%{"id" => id1}, %{"id" => id2}] <|> data
+      assert_lists_equal(data, [id1, id2], &(&1["id"] == &2))
     end
 
     @tag authentication: [user_name: "non_admin", permissions: ["view_data_structure"]]
@@ -88,20 +86,31 @@ defmodule TdDdWeb.UserSearchFilterControllerTest do
            claims: %{user_id: user_id},
            domain: %{id: domain_id}
          } do
-      %{id: id1} = insert(:user_search_filter, user_id: user_id)
+      %{id: id1} = insert(:user_search_filter, scope: "data_structure", user_id: user_id)
 
       %{id: id2} =
-        insert(:user_search_filter, filters: %{"taxonomy" => [domain_id]}, is_global: true)
+        insert(:user_search_filter,
+          scope: "data_structure",
+          filters: %{"taxonomy" => [domain_id]},
+          is_global: true
+        )
 
-      insert(:user_search_filter, filters: %{"taxonomy" => [domain_id + 1]}, is_global: true)
-      insert(:user_search_filter, is_global: false)
+      insert(:user_search_filter,
+        scope: "data_structure",
+        filters: %{"taxonomy" => [domain_id + 1]},
+        is_global: true
+      )
+
+      insert(:user_search_filter, scope: "data_structure", is_global: false)
 
       assert %{"data" => data} =
                conn
-               |> get(Routes.user_search_filter_path(conn, :index_by_user))
+               |> get(Routes.user_search_filter_path(conn, :index_by_user), %{
+                 "scope" => "data_structure"
+               })
                |> json_response(:ok)
 
-      assert [%{"id" => id1}, %{"id" => id2}] <|> data
+      assert_lists_equal(data, [id1, id2], &(&1["id"] == &2))
     end
   end